add indexing errors to index, commit before adding metadata to thread pool for indexing
Before if the getIndexFields throws an error during indexing the entire commit is rolled back and a message is logged.  However the information is extremely cryptic to the user.  The new behaviour will instead index allText and add the fact that an error has occurred to the index along with the error message (error is still logged).  We can now add an admin page for finding all the metadata that were not correctly indexed.  In addition, although the indexing failed the changes the user made are not lost and some search capabilities remain for that metadata.

When a metadata is added to be indexed in the thread pool, the dbms is now committed so that there is not a race condition for indexing the metadata.  Before if the indexing started before the request returned the old metadata would be index (or in the case of create metadata it would not exist).
In the mode without multithread indexing, there was the case that the metadata could be indexed, then an exception is thrown and the database is rolled back.  In this case the index is no longer in sync with the metadata.

git-svn-id: https://geonetwork.svn.sourceforge.net/svnroot/geonetwork/trunk@8591 ff403467-3f20-0410-99e5-ff581e626a08