Performance work for metadata.show
GetRelated triggered 2 calls to getMetadata which can be expensive, especially with lots of XLinks.

This change allows metadata.show and metadata.show.embedded to cache the metadata.  (This is configured in config-metadata.xml)
The metadata is cached and GetRelated checks to see if it is cached.  If it is them it does not call getMetadata.  The results in a
3x speed improvement.

A new service is added to the metadata.show and metadata.show.embedded that clears the cache at the end of the call so the caching
does not result in too much memory consumed.

git-svn-id: https://geonetwork.svn.sourceforge.net/svnroot/geonetwork/trunk@8677 ff403467-3f20-0410-99e5-ff581e626a08